Add new enemy for medpack picker
The new enemy doesn't face player properly. While, I will try to solve
this problem. I change the map size to 50*50, disable enemy reproduce
and only left 2 different type enemy for testing.  The navmash Astar can
handle 200*200 map.

+++ TopDownShooter/Assets/Scripts/Game Engine/EnemyBasic.cs
@@ -6,7 +6,7 @@ public class EnemyBasic : MonoBehaviour {
 	public float equipped = 0;
 	public float speed = 100;
 
-	public float distanceScale = 0.3;
+	public double distanceScale = 0.3;
 
 	//power up stuff
 	int armorTimer;
